requirements.txt
pandas
python-dotenv
flask
requests
Flask-OAuthlib
Werkzeug==2.3.7
spotipy